Change the mode for RandomAccessFile from rw to rws
The mode rws is for reading and writing, as with rw, and also requires that
every update to the file's content or metadata be written synchronously to the
underlying storage device.

This can prevent the device without battery from losing security data if the
power is cut off immediately right after the lock setting is updated.

Redmine Issue#9145

Change-Id: If02d95fd54a8f37aa3003f745038d6fb4b2455e0